公会功能基本完成，但还需完善。

+++ gamedatademo/src/main/java/com/example/gamedatademo/mapper/GuildMapper.java
@@ -11,6 +11,6 @@ import com.example.gamedatademo.bean.Guild;
  */
 public interface GuildMapper {
     Integer insert(Guild guild);
-    Guild selectByGuildId(Integer GuildId);
+    Guild selectByGuildId(Long GuildId);
     Integer updateByGuildId(Guild guild);
 }
